+++ R/googleAuthR_shiny.R
@@ -64,7 +64,9 @@ gar_shiny_getAuthUrl <-
            client.id     = getOption("googleAuthR.webapp.client_id"),
            client.secret = getOption("googleAuthR.webapp.client_secret"),
            scope         = getOption("googleAuthR.scopes.selected"),
-           access_type   = "online") {
+           access_type   = c("online","offline")) {
+    
+    access_type <- match.arg(access_type)
 
     scopeEnc <- paste(scope, sep='', collapse=' ')
     
@@ -157,7 +159,7 @@ gar_shiny_getToken <- function(code,
   # content of req will contain access_token, token_type, expires_in
   token <- httr::content(req, type = "application/json")
   
-  # Create a Token2.0 object consistent with the token obtained from scr_auth()
+  # Create a Token2.0 object consistent with the token obtained from gar_auth()
   token_formatted <-
     httr::Token2.0$new(app = gar_app,
                        endpoint = httr::oauth_endpoints("google"),
@@ -251,7 +253,6 @@ reactiveAccessToken <- function(session){
   shiny::reactive({
     ## gets all the parameters in the URL. The auth code should be one of them.
     
-    # if(length(pars$code) > 0) {
     if(!is.null(authReturnCode(session))){
       ## extract the authorization token
       app_url <- gar_shiny_getUrl(session)    
@@ -360,6 +361,7 @@ loginOutput <- function(output_name){
 #' @param logout_text What the logout text will read on the button
 #' @param login_class The Bootstrap class for the login link
 #' @param logout_class The Bootstrap class for the logout link
+#' @param revoke If TRUE a user on logout will need to re-authenticate.
 #' 
 #' @return An object to assign to output e.g. output$login
 #' 
@@ -436,7 +438,9 @@ renderLogin <- function(session,
                         logout_text="Logout",
                         login_class="btn btn-primary",
                         logout_class="btn btn-default",
-                        access_type = "online"){
+                        access_type = c("online","offline"),
+                        revoke = TRUE){
+  access_type <- match.arg(access_type)
   shiny::renderUI({
     if(is.null(shiny::isolate(access_token))) {
       shiny::actionLink("signed_in",
@@ -446,8 +450,18 @@ renderLogin <- function(session,
                    class=login_class, 
                    role="button"))
     } else {
+      if(revoke && !is.null(shiny::isolate(access_token))){
+        message("Revoked access")
+        
+        logout_url <- httr::modify_url("https://accounts.google.com/o/oauth2/revoke",
+                                       query = list(token = shiny::isolate(access_token)$credentials$access_token))
+      } else {
+        logout_url <- gar_shiny_getUrl(session)
+      }
+
+      
       shiny::a(logout_text, 
-        href = gar_shiny_getUrl(session), 
+        href = logout_url, 
         class=logout_class, 
         role="button")
     }