+++ NEWS.md
@@ -1,5 +1,10 @@
 # News Updates
 
+# v1.1.9003
+
+* Exposed `gar_shiny_getUrl` and the authentication type (online/offline) in `renderLogin`
+* `renderLogin` : logout now has option `revoke` to revoke authentication token
+
 # v1.1.9002
 
 * Added option for `googleAuthR.jsonlite.simplifyVector` for content parsing for compatibility for some APIs